==> 1_CreacionDeDatasets.ipynb <==
<h1> Creación de Datasets

<h3> Creación de arreglos de NumPy para almacenar los datos de entrenamiento, validación y prueba de ambos modelos

Equipo de Reto 5

Inteligencia artificial avanzada para la ciencia de datos II (Grupo 502)

30 de noviembre de 2023

### Preparar el entorno de trabajo

# Importar librerias y módulos necesarios
import os
import pandas as pd
import cv2
import numpy as np
import sys
from scipy.ndimage import gaussian_filter
from google.colab.patches import cv2_imshow

# Definir el directorio de trabajo actual
os.chdir('/content/drive/Shareddrives/Penta Tech/Reto')

# Definir dimensiones importantes de los datos
img_height = 112
img_width = 112
img_channels = 1
mask_channels = 1
landmarks_channels = 7

### Cargar la lista de archivos y de trazos

# Lista de archivos
file_list = pd.read_csv(os.path.join(os.getcwd(), 'EchoNet-Dynamic/FileList.csv'))
file_list.head()

# Lista de trazos
volume_tracings = pd.read_csv(os.path.join(os.getcwd(), 'EchoNet-Dynamic/VolumeTracings.csv'))
volume_tracings.head()

# Identificar todos los fotogramas trazados
labeled_frames = volume_tracings.groupby(['FileName', 'Frame']).size().reset_index(name='Count')
labeled_frames.head()

# Borrar del listado los fotogramas de videos que no existen o producen errores
labeled_frames.drop([7761, 8772, 10670, 10671, 11325, 12282, 12509, 18963, 19239, 19699], inplace=True)

### Crear arreglos para almacenar los datos de entrenamiento, validación y prueba

# Conjunto de entrenamiento
train_images = np.empty((0, img_height, img_width, img_channels), dtype=np.float16)
train_masks = np.empty((0, img_height, img_width, mask_channels), dtype=np.float16)
train_landmarks = np.empty((0, img_height, img_width, landmarks_channels), dtype=np.float16)

# Conjunto de validación
val_images = np.empty((0, img_height, img_width, img_channels), dtype=np.float16)
val_masks = np.empty((0, img_height, img_width, mask_channels), dtype=np.float16)
val_landmarks = np.empty((0, img_height, img_width, landmarks_channels), dtype=np.float16)

# Conjunto de prueba
test_images = np.empty((0, img_height, img_width, img_channels), dtype=np.float16)
test_masks = np.empty((0, img_height, img_width, mask_channels), dtype=np.float16)
test_landmarks = np.empty((0, img_height, img_width, landmarks_channels), dtype=np.float16)

### Dividir y guardar los datos de los conjuntos de entrenamiento, validación y prueba

# OPCIONAL: Seleccionar un rango de datos
labeled_frames = labeled_frames.iloc[:(labeled_frames.shape[0]//4)]

# Para todos los fotogramas trazados
for index, series in labeled_frames.iterrows():

  # Abrir el fotograma
  videos_path = os.path.join(os.getcwd(), 'EchoNet-Dynamic/Videos', series[0])
  cap = cv2.VideoCapture(videos_path)
  cap.set(cv2.CAP_PROP_POS_FRAMES, series[1])
  _, frame = cap.read()
  cap.release()

  # Ordenar los puntos del trazo utilizando el algoritmo Convex Hull
  points = volume_tracings[(volume_tracings['FileName'] == series[0]) & (volume_tracings['Frame'] == series[1])]
  c_h_points = pd.DataFrame({'X': pd.concat([points.X1, points.X2], ignore_index=True), 'Y': pd.concat([points.Y1, points.Y2], ignore_index=True)})
  c_h_points = np.array(c_h_points).astype(np.int32).reshape((-1, 1, 2))

  # Crear una máscara del fotograma
  h, w, _ = frame.shape
  mask = np.zeros((frame.shape[0], frame.shape[1], mask_channels)).astype(np.uint8)
  cv2.fillPoly(mask, [cv2.convexHull(c_h_points)], 255)

  # Almacenar las pendientes y puntos de las líneas trazadas
  slopes, left, right = [], [], []
  for _, line in points.iterrows():
    if line['X2'] == line['X1']:
      slopes.append(sys.maxsize)
      left.append((line['X1'], line['Y1']))
      right.append((line['X2'], line['Y2']))
    elif line['X2'] > line['X1']:
      slopes.append((line['Y2'] - line['Y1']) / (line['X2'] - line['X1']))
      left.append((line['X1'], line['Y1']))
      right.append((line['X2'], line['Y2']))
    else:
      slopes.append((line['Y1'] - line['Y2']) / (line['X1'] - line['X2']))
      left.append((line['X2'], line['Y2']))
      right.append((line['X1'], line['Y1']))

  # Identificar la única línea vertical trazada
  if abs(max(slopes) - np.mean(slopes)) > abs(min(slopes) - np.mean(slopes)):
    a_index = slopes.index(max(slopes))
  else:
    a_index = slopes.index(min(slopes))

  # Guardar la coordenada del punto más alto del trazo
  coords = []
  apex = points.iloc[a_index]
  if apex['Y2'] < apex['Y1']:
    coords.append((apex['X2'], apex['Y2']))
  else:
    coords.append((apex['X1'], apex['Y1']))

  # Guardar la coordenada de los puntos más bajos del trazo
  max_area = 0
  for l in left:
    for r in right:
      area = 0.5 * abs(coords[0][0] * (l[1] - r[1]) + l[0] * (r[1] - coords[0][1]) + r[0] * (coords[0][1] - l[1]))
      if area >= max_area:
        max_area = area
        bases = (l, r)
  coords.append(bases[0])
  coords.append(bases[1])

  # Identificar los puntos intermedios en cada lado
  left = sorted(left, key=lambda l: l[1])
  right = sorted(right, key=lambda r: r[1])
  left = left[:left.index(coords[1])]
  right = right[:right.index(coords[2])]
  coords.append(left[len(left)//3])
  coords.append(right[len(right)//3])
  coords.append(left[2*len(left)//3])
  coords.append(right[2*len(right)//3])

  # Crear un arreglo con los puntos más importantes
  landmarks = np.zeros((frame.shape[0], frame.shape[1], len(coords)), dtype=np.uint8)
  for coord in coords:
    channel_image = np.zeros((frame.shape[:2]))
    cv2.circle(channel_image, (int(coord[0]), int(coord[1])), 4, 255, -1)
    channel_image = gaussian_filter(channel_image, sigma=2)
    landmarks[:,:,coords.index(coord)] = channel_image

  # Guardar el fotograma, la máscara y los puntos más importantes en los arreglos correspondientes
  split = file_list.loc[file_list['FileName'] == series[0][:-4], 'Split'].iloc[0]
  greyscale_frame = np.expand_dims(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY), axis=-1).astype(np.uint8)
  if split == 'TRAIN':
      train_images = np.append(train_images.astype(np.float16), [(greyscale_frame/255.0).astype(np.float16)], axis=0)
      train_masks = np.append(train_masks.astype(np.float16), [(mask/255.0).astype(np.float16)], axis=0)
      train_landmarks = np.append(train_landmarks.astype(np.float16), [(landmarks/255.0).astype(np.float16)], axis=0)
  if split == 'VAL':
      val_images = np.append(val_images.astype(np.float16), [(greyscale_frame/255.0).astype(np.float16)], axis=0)
      val_masks = np.append(val_masks.astype(np.float16), [(mask/255.0).astype(np.float16)], axis=0)
      val_landmarks = np.append(val_landmarks.astype(np.float16), [(landmarks/255.0).astype(np.float16)], axis=0)
  if split == 'TEST':
      test_images = np.append(test_images.astype(np.float16), [(greyscale_frame/255.0).astype(np.float16)], axis=0)
      test_masks = np.append(test_masks.astype(np.float16), [(mask/255.0).astype(np.float16)], axis=0)
      test_landmarks = np.append(test_landmarks.astype(np.float16), [(landmarks/255.0).astype(np.float16)], axis=0)

  # Imprimir la información de los datos
  print(f'ÍNDICE: {index}  VIDEO: {series[0]}  FOTOGRAMA: {series[1]}  CONJUNTO: {split}')

  # Mostrar el fotograma, la máscara, los puntos más importantes y las líneas del trazo
  frame[:,:,0] = cv2.addWeighted(mask.astype(np.uint8), 0.75, frame[:,:,0], 1, 0)
  for coord in coords:
    cv2.circle(frame, (int(coord[0]), int(coord[1])), 1, (0, 0, 255), -1)
  for _, point in points.iterrows():
    cv2.line(frame, (int(point.X1), int(point.Y1)), (int(point.X2), int(point.Y2)), (0, 255, 0), thickness=1)
  cv2_imshow(frame)

# Crear directorio para guardar todos los arreglos como archivos
array_directory = os.path.join(os.getcwd(), 'Datasets')
os.mkdir(array_directory)

# Guardar todos los arreglos de entrenamiento
np.save(os.path.join(array_directory, 'train_images.npy'), train_images.astype(np.float16))
np.save(os.path.join(array_directory, 'train_masks.npy'), train_masks.astype(np.float16))
np.save(os.path.join(array_directory, 'train_landmarks.npy'), train_landmarks.astype(np.float16))

# Guardar todos los arreglos de validación
np.save(os.path.join(array_directory, 'val_images.npy'), val_images.astype(np.float16))
np.save(os.path.join(array_directory, 'val_masks.npy'), val_masks.astype(np.float16))
np.save(os.path.join(array_directory, 'val_landmarks.npy'), val_landmarks.astype(np.float16))

# Guardar todos los arreglos de prueba
np.save(os.path.join(array_directory, 'test_images.npy'), test_images.astype(np.float16))
np.save(os.path.join(array_directory, 'test_masks.npy'), test_masks.astype(np.float16))
np.save(os.path.join(array_directory, 'test_landmarks.npy'), test_landmarks.astype(np.float16))

### Validar que los arreglos se hayan guardado correctamente

# Imprimir las dimensiones de los archivos de entrenamiento
print('Imágenes de entrenamiento:', np.load(os.path.join(array_directory, 'train_images.npy')).shape)
print('Máscaras de entrenamiento:', np.load(os.path.join(array_directory, 'train_masks.npy')).shape)
print('Puntos de entrenamiento:  ', np.load(os.path.join(array_directory, 'train_landmarks.npy')).shape, '\n')

# Imprimir las dimensiones de los archvios de validación
print('Imágenes de validación:   ', np.load(os.path.join(array_directory, 'val_images.npy')).shape)
print('Máscaras de validación:   ', np.load(os.path.join(array_directory, 'val_masks.npy')).shape)
print('Puntos de validación:     ', np.load(os.path.join(array_directory, 'val_landmarks.npy')).shape, '\n')

# Imprimir las dimensiones de los archivos de prueba
print('Imágenes de prueba:       ', np.load(os.path.join(array_directory, 'test_images.npy')).shape)
print('Máscaras de prueba:       ', np.load(os.path.join(array_directory, 'test_masks.npy')).shape)
print('Puntos de prueba:         ', np.load(os.path.join(array_directory, 'test_landmarks.npy')).shape)

### OPCIONAL: Concatenar los nuevos arreglos a aquellos previamente almacenados



# Definir el directorio con los archivos generados anteriormente
complete_array_directory = os.path.join(os.getcwd(), 'CompleteDatasets')

# Añadir los nuevos datos a los conjuntos existentes
for file in os.listdir(complete_array_directory):
  array = np.load(os.path.join(array_directory, file))
  complete_array = np.load(os.path.join(complete_array_directory, file))
  complete_array = np.concatenate((array, complete_array), axis=0)
  np.save(os.path.join(complete_array_directory, file), complete_array)
  print(f'El archivo {file} se actualizó con éxito')
  print(f'  Dimensiones: {complete_array.shape}  Tipo: {complete_array.dtype}')
